requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
python-multipart==0.0.6
langchain==0.0.312
pyttsx3==2.90
llama-cpp-python==0.2.10
transformers==4.35.0
torch==2.1.0
numpy==1.26.0
aiofiles==23.2.1
